optimize: abstract undo update executor
