+++ rm-datasource/src/main/java/io/seata/rm/datasource/undo/oracle/OracleUndoUpdateExecutor.java
@@ -15,55 +15,15 @@
  */
 package io.seata.rm.datasource.undo.oracle;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
-import io.seata.common.exception.ShouldNeverHappenException;
-import io.seata.common.util.CollectionUtils;
-import io.seata.rm.datasource.ColumnUtils;
-import io.seata.rm.datasource.SqlGenerateUtils;
-import io.seata.rm.datasource.sql.struct.Field;
-import io.seata.rm.datasource.sql.struct.Row;
-import io.seata.rm.datasource.sql.struct.TableRecords;
-import io.seata.rm.datasource.undo.AbstractUndoExecutor;
+import io.seata.rm.datasource.undo.AbstractUndoUpdateExecutor;
 import io.seata.rm.datasource.undo.SQLUndoLog;
-import io.seata.sqlparser.util.JdbcConstants;
 
 /**
  * The type oracle undo update executor.
  *
  * @author ccg
  */
-public class OracleUndoUpdateExecutor extends AbstractUndoExecutor {
-
-    /**
-     * UPDATE a SET x = ?, y = ?, z = ? WHERE pk1 = ? and pk2 = ?
-     */
-    private static final String UPDATE_SQL_TEMPLATE = "UPDATE %s SET %s WHERE %s ";
-
-    @Override
-    protected String buildUndoSQL() {
-        TableRecords beforeImage = sqlUndoLog.getBeforeImage();
-        List<Row> beforeImageRows = beforeImage.getRows();
-        if (CollectionUtils.isEmpty(beforeImageRows)) {
-            throw new ShouldNeverHappenException("Invalid UNDO LOG"); // TODO
-        }
-        Row row = beforeImageRows.get(0);
-
-        List<Field> nonPkFields = row.nonPrimaryKeys();
-        // update sql undo log before image all field come from table meta. need add escape.
-        // see BaseTransactionalExecutor#buildTableRecords
-        String updateColumns = nonPkFields.stream().map(
-            field -> ColumnUtils.addEscape(field.getName(), JdbcConstants.ORACLE) + " = ?").collect(
-            Collectors.joining(", "));
-
-        List<String> pkNameList = getOrderedPkList(beforeImage, row, JdbcConstants.ORACLE).stream().map(
-            e -> e.getName()).collect(Collectors.toList());
-        String whereSql = SqlGenerateUtils.buildWhereConditionByPKs(pkNameList, JdbcConstants.ORACLE);
-
-        return String.format(UPDATE_SQL_TEMPLATE, sqlUndoLog.getTableName(), updateColumns, whereSql);
-    }
-
+public class OracleUndoUpdateExecutor extends AbstractUndoUpdateExecutor {
     /**
      * Instantiates a new My sql undo update executor.
      *
@@ -71,10 +31,5 @@ public class OracleUndoUpdateExecutor extends AbstractUndoExecutor {
      */
     public OracleUndoUpdateExecutor(SQLUndoLog sqlUndoLog) {
         super(sqlUndoLog);
-    }
-
-    @Override
-    protected TableRecords getUndoRows() {
-        return sqlUndoLog.getBeforeImage();
     }
 }
